requirements.txt
django
pillow
pandas
matplotlib
openai
numpy
requests
python-dotenv